xfz/requirements.txt
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-dysmsapi==1.0.0
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
Django==2.0
django-debug-toolbar==2.2
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
idna==2.10
jieba==0.42.1
jmespath==0.10.0
mysqlclient==1.4.6
packaging==20.4
Pillow==7.2.0
PyMySQL==0.9.3
pyparsing==2.4.7
python-memcached==1.59
pytz==2020.1
qiniu==7.2.8
redis==3.5.3
requests==2.24.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
webencodings==0.5.1
Whoosh==2.7.4
